requirements.txt
pytest==8.3.3
Appium-Python-Client==4.2.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ffmpeg==1.4
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.1.2
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
Pillow-PIL==0.1.dev0
pluggy==1.5.0
psutil==6.0.0
PySocks==1.7.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-runner==6.0.1
pytest-variables==3.1.0
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
